+++ src/DataSets/DataSets.component.js
@@ -39,7 +39,7 @@ import FlatButton from 'material-ui/FlatButton/FlatButton';
 import HelpOutlineIcon from 'material-ui/svg-icons/action/help-outline';
 import ListIcon from 'material-ui/svg-icons/action/list';
 import FormHelpers from '../forms/FormHelpers';
-import {currentUserHasAdminRole} from '../utils/Dhis2Helpers';
+import {currentUserHasAdminRole, canCreate} from '../utils/Dhis2Helpers';
 
 const {SimpleCheckBox} = FormHelpers;
 
@@ -296,7 +296,7 @@ const DataSets = React.createClass({
         );
 
         const {d2} = this.context;
-        const userCanCreateDataSets = d2.currentUser.canCreatePrivate(d2.models.dataSet);
+        const userCanCreateDataSets = canCreate(d2);
 
         const logActions = [
             <FlatButton
